Extract FeedLoderStub into a shared scope to remove duplication
