+++ EssentialApp/EssentialAppTests/FeedLoaderWithFallbackCompositeTests.swift
@@ -40,8 +40,8 @@ final class FeedLoaderWithFallbackCompositeTests: XCTestCase {
         file: StaticString = #file,
         line: UInt = #line
     ) -> FeedLoader {
-        let primaryLoader = LoaderStub(result: primaryResult)
-        let fallbackLoader = LoaderStub(result: fallbackResult)
+        let primaryLoader = FeedLoaderStub(result: primaryResult)
+        let fallbackLoader = FeedLoaderStub(result: fallbackResult)
         let sut = FeedLoaderWithFallbackComposite(primary: primaryLoader, fallback: fallbackLoader)
 
         trackMemoryLeaks(primaryLoader)
@@ -79,19 +79,6 @@ final class FeedLoaderWithFallbackCompositeTests: XCTestCase {
     private func uniqueFeed() -> [FeedImage] {
         [FeedImage(id: UUID(), description: "any", location: "any", url: URL(string: "http:\\any-url.com")!)]
     }
-
-    private class LoaderStub: FeedLoader {
-        private let result: FeedLoader.Result
-
-        init(result: FeedLoader.Result) {
-            self.result = result
-        }
-
-        func load(completion: @escaping (FeedLoader.Result) -> (Void)) {
-            completion(result)
-        }
-    }
-
 }
